+++ src/modules/aboutDialog.js
@@ -1,6 +1,6 @@
 // aboutDialog.js
 
-const VERSION = "1.1.0";
+const VERSION = "1.2.0";
 
 export class AboutDialog {
   constructor(i18n) {